project/requirements.txt
asgiref==3.6.0
astroid==2.15.3
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2023.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0

celery==5.1.2
redis==3.5.3
